Module pallet_collator_selection::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub use crate::weights::WeightInfo;

pub use __tt_default_parts_6 as tt_default_parts;

Structs

CandidateInfo

Basic information about a collation candidate.

GenesisConfig

Can be used to configure the genesis state of this pallet.

IdentityCollator

A convertor from collators id. Since this pallet does not have stash/controller, this is just identity.

Pallet

The pallet implementing the on-chain logic.

_GeneratedPrefixForStorageCandidacyBond

_GeneratedPrefixForStorageCandidates

_GeneratedPrefixForStorageDesiredCandidates

_GeneratedPrefixForStorageInvulnerables

_GeneratedPrefixForStorageLastAuthoredBlock

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configure the pallet by specifying the parameters and types on which it depends.

Type Definitions

CandidacyBond

Fixed deposit bond for each candidate.

Candidates

The (community, limited) collation candidates.

DesiredCandidates

Desired number of candidates.

Invulnerables

The invulnerable, fixed collators.

LastAuthoredBlock

Last block authored by collator.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.