Function pallet_collator_selection::pallet::dispatchables::update_bond

pub fn update_bond<T: Config>(
    new_deposit: <<T as Config>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance
)

Update the candidacy bond of collator candidate origin to a new amount new_deposit.

Setting a new_deposit that is lower than the current deposit while origin is occupying a top-DesiredCandidates slot is not allowed.

This call will fail if origin is not a collator candidate, the updated bond is lower than the minimum candidacy bond, and/or the amount cannot be reserved.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::update_bond.