pallet_collator_selection::pallet

Trait AccountCheck

pub trait AccountCheck<AccountId> {
    // Required method
    fn allowed_candidacy(account: &AccountId) -> bool;
}

Used to check whether an account is allowed to be a candidate.

Required Methods

fn allowed_candidacy(account: &AccountId) -> bool

true if the account is allowed to be a candidate, false otherwise.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.