Trait pallet_collator_selection::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Currency: LockableCurrency<Self::AccountId>;
    type UpdateOrigin: EnsureOrigin<Self::Origin>;
    type PotId: Get<PalletId>;
    type MaxCandidates: Get<u32>;
    type MinCandidates: Get<u32>;
    type MaxInvulnerables: Get<u32>;
    type KickThreshold: Get<Self::BlockNumber>;
    type ValidatorId: Member + Parameter;
    type ValidatorIdOf: Convert<Self::AccountId, Option<Self::ValidatorId>>;
    type ValidatorRegistration: ValidatorRegistration<Self::ValidatorId>;
    type WeightInfo: WeightInfo;
}

Configure the pallet by specifying the parameters and types on which it depends.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Overarching event type.

type Currency: LockableCurrency<Self::AccountId>

The currency mechanism.

type UpdateOrigin: EnsureOrigin<Self::Origin>

Origin that can dictate updating parameters of this pallet.

type PotId: Get<PalletId>

Account Identifier from which the internal Pot is generated.

type MaxCandidates: Get<u32>

Maximum number of candidates that we should have. This is used for benchmarking and is not enforced.

This does not take into account the invulnerables.

type MinCandidates: Get<u32>

Minimum number of candidates that we should have. This is used for disaster recovery.

This does not take into account the invulnerables.

type MaxInvulnerables: Get<u32>

Maximum number of invulnerables.

Used only for benchmarking.

type KickThreshold: Get<Self::BlockNumber>

type ValidatorId: Member + Parameter

A stable ID for a validator.

type ValidatorIdOf: Convert<Self::AccountId, Option<Self::ValidatorId>>

A conversion from account ID to validator ID.

Its cost must be at most one storage read.

type ValidatorRegistration: ValidatorRegistration<Self::ValidatorId>

Validate a user is registered

type WeightInfo: WeightInfo

The weight information of this pallet.