Type Alias pallet_collator_selection::pallet::Module

pub type Module<T> = Pallet<T>;

👎Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

pub fn set_invulnerables( origin: OriginFor<T>, new: Vec<T::AccountId> ) -> DispatchResult

Set the list of invulnerable (fixed) collators. These collators must do some preparation, namely to have registered session keys.

The call will remove any accounts that have not registered keys from the set. That is, it is non-atomic; the caller accepts all AccountIds passed in new individually as acceptable Invulnerables, and is not proposing a set of new Invulnerables.

This call does not maintain mutual exclusivity of Invulnerables and Candidates. It is recommended to use a batch of add_invulnerable and remove_invulnerable instead. A batch_all can also be used to enforce atomicity. If any candidates are included in new, they should be removed with remove_invulnerable_candidate after execution.

Must be called by the UpdateOrigin.

pub fn set_desired_candidates( origin: OriginFor<T>, max: u32 ) -> DispatchResultWithPostInfo

Set the ideal number of non-invulnerable collators. If lowering this number, then the number of running collators could be higher than this figure. Aside from that edge case, there should be no other way to have more candidates than the desired number.

The origin for this call must be the UpdateOrigin.

pub fn set_candidacy_bond( origin: OriginFor<T>, bond: <<T as Config>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance ) -> DispatchResultWithPostInfo

Set the candidacy bond amount.

If the candidacy bond is increased by this call, all current candidates which have a deposit lower than the new bond will be kicked from the list and get their deposits back.

The origin for this call must be the UpdateOrigin.

pub fn register_as_candidate(origin: OriginFor<T>) -> DispatchResultWithPostInfo

Register this account as a collator candidate. The account must (a) already have registered session keys and (b) be able to reserve the CandidacyBond.

This call is not available to Invulnerable collators.

pub fn leave_intent(origin: OriginFor<T>) -> DispatchResultWithPostInfo

Deregister origin as a collator candidate. Note that the collator can only leave on session change. The CandidacyBond will be unreserved immediately.

This call will fail if the total number of candidates would drop below MinEligibleCollators.

pub fn add_invulnerable( origin: OriginFor<T>, who: T::AccountId ) -> DispatchResultWithPostInfo

Add a new account who to the list of Invulnerables collators. who must have registered session keys. If who is a candidate, they will be removed.

The origin for this call must be the UpdateOrigin.

pub fn remove_invulnerable( origin: OriginFor<T>, who: T::AccountId ) -> DispatchResult

Remove an account who from the list of Invulnerables collators. Invulnerables must be sorted.

The origin for this call must be the UpdateOrigin.

pub fn update_bond( origin: OriginFor<T>, new_deposit: <<T as Config>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance ) -> DispatchResultWithPostInfo

Update the candidacy bond of collator candidate origin to a new amount new_deposit.

Setting a new_deposit that is lower than the current deposit while origin is occupying a top-DesiredCandidates slot is not allowed.

This call will fail if origin is not a collator candidate, the updated bond is lower than the minimum candidacy bond, and/or the amount cannot be reserved.

pub fn take_candidate_slot( origin: OriginFor<T>, deposit: <<T as Config>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance, target: T::AccountId ) -> DispatchResultWithPostInfo

The caller origin replaces a candidate target in the collator candidate list by reserving deposit. The amount deposit reserved by the caller must be greater than the existing bond of the target it is trying to replace.

This call will fail if the caller is already a collator candidate or invulnerable, the caller does not have registered session keys, the target is not a collator candidate, and/or the deposit amount cannot be reserved.

impl<T: Config> Pallet<T>

pub fn account_id() -> T::AccountId

Get a unique, inaccessible account ID from the PotId.

pub fn assemble_collators() -> Vec<T::AccountId>

Assemble the current set of candidates and invulnerables into the next collator set.

This is done on the fly, as frequent as we are told to do so, as the session manager.

pub fn kick_stale_candidates( candidates: impl IntoIterator<Item = T::AccountId> ) -> u32

Kicks out candidates that did not produce a block in the kick threshold and refunds their deposits.

Return value is the number of candidates left in the list.

pub fn do_try_state() -> Result<(), TryRuntimeError>

Ensure the correctness of the state of this pallet.

This should be valid before or after each state transition of this pallet.

Invariants

DesiredCandidates

The current desired candidate count should not exceed the candidate list capacity.

The number of selected candidates together with the invulnerables must be greater than or equal to the minimum number of eligible collators.

impl<T: Config> Pallet<T>

pub fn invulnerables() -> BoundedVec<T::AccountId, T::MaxInvulnerables>

An auto-generated getter for Invulnerables.

impl<T: Config> Pallet<T>

pub fn candidate_list( ) -> BoundedVec<CandidateInfo<T::AccountId, <<T as Config>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance>, T::MaxCandidates>

An auto-generated getter for CandidateList.

impl<T: Config> Pallet<T>

pub fn last_authored_block<KArg>(k: KArg) -> BlockNumberFor<T>where KArg: EncodeLike<T::AccountId>,

An auto-generated getter for LastAuthoredBlock.

impl<T: Config> Pallet<T>

pub fn desired_candidates() -> u32

An auto-generated getter for DesiredCandidates.

impl<T: Config> Pallet<T>

pub fn candidacy_bond( ) -> <<T as Config>::Currency as Currency<<T as SystemConfig>::AccountId>>::Balance

An auto-generated getter for CandidacyBond.