pallet_collator_selection

Module pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

The main components of this pallet are:

Pallet, which implements all of the dispatchable extrinsics of the pallet, among other public functions.

The subset of the functions that are dispatchable can be identified either in the dispatchables module or in the Call enum.

storage_types, which contains the list of all types that are representing a storage item. Otherwise, all storage items are listed among Type Definitions.

Config, which contains the configuration trait of this pallet.

Event and Error, which are listed among the Enums.

Re-exports

pub use crate::weights::WeightInfo;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Structs

CandidateInfo

Basic information about a collation candidate.

GenesisConfig

Can be used to configure the genesis state of this pallet.

IdentityCollator

A convertor from collators id. Since this pallet does not have stash/controller, this is just identity.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

Traits

AccountCheck

Used to check whether an account is allowed to be a candidate.

Config

Configuration trait of this pallet.

Type Aliases

CandidacyBond

Fixed amount to deposit to become a collator.

Candidates

The (community approved, limited) collation candidates.

DesiredCandidates

Desired number of candidates.

Invulnerables

The invulnerable, fixed collators.

LastAuthoredBlock

Last block authored by collator.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

NonCandidates

Candidates who initiated leave intent or kicked.

PendingApplications

Pending applications to become a collator.

SlashDestination

Destination account for slashed amount.