Module pallet_collator_selection::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

leave_intent

Deregister origin as a collator candidate. Note that the collator can only leave on session change. The CandidacyBond will start un-bonding process.

register_as_candidate

Register this account as a collator candidate. The account must (a) already have registered session keys and (b) be able to reserve the CandidacyBond.

set_candidacy_bond

Set the candidacy bond amount.

set_desired_candidates

Set the ideal number of collators (not including the invulnerables). If lowering this number, then the number of running collators could be higher than this figure. Aside from that edge case, there should be no other way to have more collators than the desired number.

set_invulnerables

Set the list of invulnerable (fixed) collators.

set_slash_destination

Set slash destination. Use Some to deposit slashed balance into destination or None to burn it.

withdraw_bond

Withdraw CandidacyBond after un-bonding period has finished. This call will fail called during un-bonding or if there’s no CandidacyBound reserved.