Module pallet_collator_selection::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

add_invulnerable

Add a new account who to the list of Invulnerables collators. who must have registered session keys. If who is a candidate, they will be removed.

leave_intent

Deregister origin as a collator candidate. Note that the collator can only leave on session change. The CandidacyBond will be unreserved immediately.

register_as_candidate

Register this account as a collator candidate. The account must (a) already have registered session keys and (b) be able to reserve the CandidacyBond.

remove_invulnerable

Remove an account who from the list of Invulnerables collators. Invulnerables must be sorted.

set_candidacy_bond

Set the candidacy bond amount.

set_desired_candidates

Set the ideal number of non-invulnerable collators. If lowering this number, then the number of running collators could be higher than this figure. Aside from that edge case, there should be no other way to have more candidates than the desired number.

set_invulnerables

Set the list of invulnerable (fixed) collators. These collators must do some preparation, namely to have registered session keys.