Function pallet_collator_selection::pallet::dispatchables::set_invulnerables

pub fn set_invulnerables<T: Config>(new: Vec<T::AccountId>)

Set the list of invulnerable (fixed) collators. These collators must do some preparation, namely to have registered session keys.

The call will remove any accounts that have not registered keys from the set. That is, it is non-atomic; the caller accepts all AccountIds passed in new individually as acceptable Invulnerables, and is not proposing a set of new Invulnerables.

This call does not maintain mutual exclusivity of Invulnerables and Candidates. It is recommended to use a batch of add_invulnerable and remove_invulnerable instead. A batch_all can also be used to enforce atomicity. If any candidates are included in new, they should be removed with remove_invulnerable_candidate after execution.

Must be called by the UpdateOrigin.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::set_invulnerables.